// src/app/models/user-info.model.ts
import { UserProfile } from './user-profile.model'

export class UserInfo {
    adresse: string;
    cabinet: string;
    codePostal: string;
    email: string;
    etat: string;
    login: string;
    matriculeFiscal: string;
    nom: string;
    password: string;
    prenom: string;
    profil: UserProfile;
    sexe: string;
    telFixe: string;
    telPortable: string;
    type: string;
    ville: string;
}